refactor: rebase auth-6
