+++ api/src/main/java/com/tago/api/auth/presentation/AuthApi.java
@@ -3,16 +3,14 @@ package com.tago.api.auth.presentation;
 
 import com.tago.api.auth.application.OAuthLoginService;
 import com.tago.api.auth.application.OAuthSdkService;
-import com.tago.api.auth.dto.LoginRequest;
-import com.tago.api.auth.dto.LoginResponse;
-import com.tago.api.auth.infra.kakao.KakaoLoginParams;
+import com.tago.api.auth.dto.request.LoginRequest;
+import com.tago.api.auth.application.TokenReissueService;
+import com.tago.api.auth.dto.response.LoginResponse;
+import com.tago.api.auth.dto.response.TokenReissueResponse;
 import com.tago.api.common.model.dto.ResponseDto;
 import lombok.RequiredArgsConstructor;
 import org.springframework.http.ResponseEntity;
-import org.springframework.web.bind.annotation.PostMapping;
-import org.springframework.web.bind.annotation.RequestBody;
-import org.springframework.web.bind.annotation.RequestMapping;
-import org.springframework.web.bind.annotation.RestController;
+import org.springframework.web.bind.annotation.*;
 
 @RestController
 @RequiredArgsConstructor
@@ -20,15 +18,17 @@ import org.springframework.web.bind.annotation.RestController;
 public class AuthApi {
     private final OAuthLoginService oAuthLoginService;
     private final OAuthSdkService oAuthSdkService;
-
-    @PostMapping("/auth/kakao")
-    public ResponseEntity<LoginResponse> loginKakao(@RequestBody KakaoLoginParams params) {
-        return ResponseDto.ok(oAuthLoginService.login(params));
-    }
+    private final TokenReissueService tokenReissueService;
 
     @PostMapping("/auth/login")
     public ResponseEntity<LoginResponse> login(@RequestBody LoginRequest loginRequest){
         return ResponseDto.ok(oAuthSdkService.login(loginRequest));
     }
-
+    @PostMapping("/auth/token/reissue")
+    public ResponseEntity<TokenReissueResponse> reissue(
+            @RequestHeader("Authorization") String refreshToken
+    ) {
+        TokenReissueResponse response = tokenReissueService.reissue(refreshToken);
+        return ResponseDto.ok(response);
+    }
 }
